add clickhouse server
