+++ marimo/_dependencies/dependencies.py
@@ -169,6 +169,7 @@ class DependencyManager:
     altair = Dependency("altair", min_version="5.3.0", max_version="6.0.0")
     duckdb = Dependency("duckdb")
     chdb = Dependency("chdb")
+    clickhouse_connect = Dependency("clickhouse_connect")
     sqlglot = Dependency("sqlglot")
     pillow = Dependency("PIL")
     plotly = Dependency("plotly")
