+++ marimo/_sql/engines.py
@@ -15,28 +15,18 @@ from marimo._data.models import (
 )
 from marimo._dependencies.dependencies import DependencyManager
 from marimo._sql.types import SQLEngine
-from marimo._sql.utils import wrapped_sql
+from marimo._sql.utils import raise_df_import_error, wrapped_sql
 from marimo._types.ids import VariableName
 
 LOGGER = _loggers.marimo_logger()
 
 if TYPE_CHECKING:
-    import chdb
     import duckdb
     from sqlalchemy import Engine
     from sqlalchemy.sql.type_api import TypeEngine
 
 # Internal engine names
 INTERNAL_DUCKDB_ENGINE = cast(VariableName, "__marimo_duckdb")
-INTERNAL_CLICKHOUSE_ENGINE = cast(VariableName, "__marimo_clickhouse")
-
-
-def raise_df_import_error(pkg: str) -> None:
-    raise ModuleNotFoundError(
-        "pandas or polars is required to execute sql. "
-        + "You can install them with 'pip install pandas polars'",
-        name=pkg,
-    )
 
 
 class DuckDBEngine(SQLEngine):
@@ -476,52 +466,6 @@ class SQLAlchemyEngine(SQLEngine):
 
     def _is_cheap_discovery(self) -> bool:
         return self.dialect.lower() in ("sqlite", "mysql", "postgresql")
-
-
-class ClickhouseEmbedded(SQLEngine):
-    """Use chdb to connect to an embedded Clickhouse"""
-
-    def __init__(
-        self,
-        connection: Optional[chdb.state.sqlitelike.Connection] = None,
-        engine_name: Optional[VariableName] = None,
-    ) -> None:
-        self._connection = connection
-        self._engine_name = engine_name
-
-    @property
-    def source(self) -> str:
-        return "clickhouse"
-
-    @property
-    def dialect(self) -> str:
-        return "clickhouse"
-
-    def execute(self, query: str) -> Any:
-        # chdb currently only supports pandas
-        if not DependencyManager.pandas.has():
-            raise_df_import_error("pandas")
-
-        import chdb
-        import pandas as pd
-
-        try:
-            result = chdb.query(query, "Dataframe")
-        except Exception:
-            LOGGER.exception("Failed to execute query")
-            return None
-        if isinstance(result, pd.DataFrame):
-            return result
-        return None
-
-    @staticmethod
-    def is_compatible(var: Any) -> bool:
-        if not DependencyManager.chdb.imported():
-            return False
-
-        from chdb.state.sqlitelike import Connection
-
-        return isinstance(var, Connection)
 
 
 def _sql_type_to_data_type(type_str: str) -> DataType:
